requirements.txt
FLAML==2.3.2
GitPython==3.1.43
Jinja2==3.1.4
MarkupSafe==2.1.5
Pygments==2.18.0
PyYAML==6.0.2
annotated-types==0.7.0
anyio==4.6.2
altair==5.4.1
attrs==24.2.0
autogen-agentchat==0.2.39
azure-cognitiveservices-speech==1.41.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lightgbm==4.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
moviepy==1.0.3
narwhals==1.9.1
numpy==2.1.2
nvidia-nccl-cu12==2.23.4
oauthlib==3.2.2
openai==1.56.2
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
streamlit==1.39.0
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchdog==5.0.3
websockets==12.0
xgboost==2.1.3